Move JsonEditor component to es_ui_shared folder
