+++ src/plugins/es_ui_shared/public/components/json_editor/use_json.ts
@@ -1,13 +1,26 @@
 /*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License;
- * you may not use this file except in compliance with the Elastic License.
+ * Licensed to Elasticsearch B.V. under one or more contributor
+ * license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright
+ * ownership. Elasticsearch B.V. licenses this file to you under
+ * the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
  */
 
 import { useEffect, useState } from 'react';
 import { i18n } from '@kbn/i18n';
 
-import { isJSON } from '../../../../../../../../src/plugins/es_ui_shared/static/validators/string';
+import { isJSON } from '../../../static/validators/string';
 
 export type OnUpdateHandler<T = { [key: string]: any }> = (arg: {
   data: {
