--- x-pack/legacy/plugins/index_management/public/app/components/json_editor/index.ts
@@ -1,9 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License;
- * you may not use this file except in compliance with the Elastic License.
- */
-
-export * from './json_editor';
-
-export { OnUpdateHandler } from './use_json';
