+++ x-pack/legacy/plugins/index_management/public/app/components/mappings_editor/components/configuration_form/meta_field_section/meta_field_section.tsx
@@ -10,8 +10,7 @@ import { FormattedMessage } from '@kbn/i18n/react';
 import { EuiLink } from '@elastic/eui';
 
 import { documentationService } from '../../../../../services/documentation';
-import { getUseField, FormRow, Field } from '../../../shared_imports';
-import { JsonEditor } from '../../../../json_editor';
+import { getUseField, FormRow, Field, JsonEditor } from '../../../shared_imports';
 
 const UseField = getUseField({ component: Field });
