+++ x-pack/legacy/plugins/index_management/public/app/components/mappings_editor/shared_imports.ts
@@ -41,3 +41,5 @@ export {
   fieldFormatters,
   fieldValidators,
 } from '../../../../../../../../src/plugins/es_ui_shared/static/forms/helpers';
+
+export { JsonEditor } from '../../../../../../../../src/plugins/es_ui_shared/public';
